requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
protobuf==5.26.0
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
